make example more specific
